+++ common/engine/keyboardprocessor/tests/unit/kmx/kmx.cpp
@@ -4,14 +4,17 @@
   Create Date:  30 Oct 2018
   Authors:      Marc Durdin (MD), Tim Eves (TSE)
 */
-#include <string>
-#include <fstream>
-#include <sstream>
-#include <cctype>
 #include <algorithm>
+#include <cctype>
+#include <fstream>
 #include <iostream>
 #include <list>
+#include <sstream>
+#include <string>
+#include <type_traits>
+
 #include <keyman/keyboardprocessor.h>
+
 #include "state.hpp"
 
 #define   try_status(expr) \
@@ -26,8 +29,6 @@ std::string utf16_to_utf8(std::u16string utf16_string); // defined in keyboard.c
 
 namespace
 {
-
-const std::string base = "tests/unit/kmx/";
 
 bool g_beep_found = false;
 
@@ -48,8 +49,8 @@ struct kmx_option {
 
 using kmx_options = std::vector<kmx_option>;
 
-int run_test(const std::string & file);
-int load_source(const std::string & file, std::string & keys, std::u16string & expected, std::u16string & context, kmx_options &options, bool &expected_beep);
+int load_source(const std::string &, std::string &, std::u16string &,
+                std::u16string &, kmx_options &, bool &);
 
 km_kbp_option_item test_env_opts[] =
 {
@@ -77,26 +78,8 @@ static inline void trim(std::string &s) {
   rtrim(s);
 }
 
-// trim from start (copying)
-static inline std::string ltrim_copy(std::string s) {
-  ltrim(s);
-  return s;
-}
-
-// trim from end (copying)
-static inline std::string rtrim_copy(std::string s) {
-  rtrim(s);
-  return s;
-}
-
-// trim from both ends (copying)
-static inline std::string trim_copy(std::string s) {
-  trim(s);
-  return s;
-}
-
 key_event char_to_event(char ch) {
-  assert(ch >= 32 && ch < 128);
+  assert(ch >= 32 && ch <= 127);
   return {
     km::kbp::kmx::s_char_to_vkey[(int)ch - 32].vk,
     (uint16_t)(km::kbp::kmx::s_char_to_vkey[(int)ch - 32].shifted ?  KM_KBP_MODIFIER_SHIFT : 0)
@@ -206,14 +189,13 @@ void apply_action(km_kbp_state const * state, km_kbp_action_item const & act, st
     }
     break;
   case KM_KBP_IT_PERSIST_OPT:
-  case KM_KBP_IT_RESET_OPT:
     assert(false); // TODO
     break;
-  case KM_KBP_IT_VKEYDOWN:
-  case KM_KBP_IT_VKEYUP:
-  case KM_KBP_IT_VSHIFTDOWN:
-  case KM_KBP_IT_VSHIFTUP:
-    assert(false); // NOT SUPPORTED
+  case KM_KBP_IT_INVALIDATE_CONTEXT:
+    std::cout << "action: context invalidated (markers cleared)" << std::endl;
+    break;
+  case KM_KBP_IT_EMIT_KEYSTROKE:
+    std::cout << "action: emit keystroke" << std::endl;
     break;
   default:
     assert(false); // NOT SUPPORTED
@@ -221,21 +203,45 @@ void apply_action(km_kbp_state const * state, km_kbp_action_item const & act, st
   }
 }
 
-int run_test(const std::string & file) {
+template<typename P>
+std::basic_string<
+  typename std::remove_const<
+    typename std::remove_pointer<P>::type>::type
+>
+utf8_to(const std::string &);
+
+template<>
+inline
+std::basic_string<wchar_t> utf8_to<const wchar_t *>(const std::string & s)
+{
+  std::wstring_convert<std::codecvt_utf8_utf16<wchar_t>> convert;
+  return convert.from_bytes(s);
+}
+
+template<>
+inline
+std::basic_string<char> utf8_to<const char *>(const std::string & s)
+{
+  return s;
+}
+
+int run_test(const std::string & source, const std::string & _compiled) {
   std::string keys = "";
   std::u16string expected = u"", context = u"";
+  auto compiled = utf8_to<km_kbp_path_name>(_compiled);
   kmx_options options;
   bool expected_beep = false;
-  
-  int result = load_source(file, keys, expected, context, options, expected_beep);
+
+  int result = load_source(source, keys, expected, context, options, expected_beep);
   if (result != 0) return result;
 
-  std::cout << "file = " << file << std::endl;
-  
+  std::cout << "source file   = " << source << std::endl
+            << "compiled file = " << _compiled << std::endl;
+
   km_kbp_keyboard * test_kb = nullptr;
   km_kbp_state * test_state = nullptr;
-    
-  try_status(km_kbp_keyboard_load(std::filesystem::path(base + file + ".kmx").c_str(), &test_kb));
+
+  try_status(km_kbp_keyboard_load(compiled.c_str(), &test_kb));
 
   // Setup state, environment
 
@@ -281,7 +287,7 @@ int run_test(const std::string & file) {
 
     try_status(km_kbp_options_update(test_state, keyboard_opts));
 
-    delete keyboard_opts;
+    delete [] keyboard_opts;
   }
 
   // Setup context
@@ -296,6 +302,7 @@ int run_test(const std::string & file) {
   // Run through key events, applying output for each event
   for (auto p = next_key(keys); p.vk != 0; p = next_key(keys)) {
     try_status(km_kbp_process_event(test_state, p.vk, p.modifier_state));
+
     for (auto act = km_kbp_state_action_items(test_state, nullptr); act->type != KM_KBP_IT_END; act++) {
       apply_action(test_state, *act, text_store);
     }
@@ -324,7 +331,7 @@ int run_test(const std::string & file) {
 
   // Test resultant options
   // TODO: test also KM_KBP_IT_PERSIST_OPT and KM_KBP_IT_RESET_OPT actions
-  
+
   for (auto it = options.begin(); it != options.end(); it++) {
     if (it->type != KOT_OUTPUT) continue;
     std::cout << "output option-key: " << utf16_to_utf8(it->key) << " expected: " << utf16_to_utf8(it->value);
@@ -349,11 +356,11 @@ std::u16string parse_source_string(std::string const & s) {
       p++;
       km_kbp_usv v;
       assert(p != s.end());
-      if (*p == 'u') {
+      if (*p == 'u' || *p == 'U') {
         // Unicode value
         p++;
         size_t n;
-        std::string s1 = s.substr(p - s.begin(), 6);
+        std::string s1 = s.substr(p - s.begin(), 8);
         v = std::stoul(s1, &n, 16);
         assert(v >= 0x20 && v <= 0x10FFFF);
         p += n-1;
@@ -381,9 +388,9 @@ std::u16string parse_source_string(std::string const & s) {
 bool parse_option_string(std::string line, kmx_options &options, kmx_option_type type) {
   auto x = line.find('=');
   if (x == std::string::npos) return false;
-  
+
   kmx_option o;
-  
+
   o.type = type;
   o.key = parse_source_string(line.substr(0, x));
   o.value = parse_source_string(line.substr(x + 1));
@@ -401,7 +408,7 @@ bool is_token(const std::string token, std::string &line) {
   return false;
 }
 
-int load_source(const std::string & file, std::string & keys, std::u16string & expected, std::u16string & context, kmx_options &options, bool &expected_beep) {
+int load_source(const std::string & path, std::string & keys, std::u16string & expected, std::u16string & context, kmx_options &options, bool &expected_beep) {
   const std::string s_keys = "c keys: ",
     s_expected = "c expected: ",
     s_context = "c context: ",
@@ -409,10 +416,7 @@ int load_source(const std::string & file, std::string & keys, std::u16string & e
     s_option_expected = "c expected option: ";
 
   // Parse out the header statements in file.kmn that tell us (a) environment, (b) key sequence, (c) start context, (d) expected result
-  std::ifstream kmn(base + file + ".kmn");
-  if (!kmn.good()) {
-    kmn.open(file + ".kmn");
-  }
+  std::ifstream kmn(path);
   if (!kmn.good()) {
     return __LINE__;
   }
@@ -452,16 +456,21 @@ int load_source(const std::string & file, std::string & keys, std::u16string & e
   return 0;
 }
 
+constexpr const auto help_str = "\
+kmx <KMN_FILE> <KMX_FILE>\n\
+help:\n\
+\tKMN_FILE:\tThe source file for the keyboard under test.\n\
+\tKMX_FILE:\tThe corresponding compiled kmx file produced from KMN_FILE.\n";
 } // namespace
 
 int main(int argc, char *argv[])
 {
-  for (int i = 1; i < argc; i++) {
-    int result;
-    std::cout << argv[i] << std::endl;
-    if ((result = run_test(argv[i])) != 0) return result;
+  if (argc < 3)
+  {
+    std::cerr << "kmx: Not enough arguments." << std::endl;
+    std::cout << help_str;
+    return 1;
   }
-  return 0;
+
+  return run_test(argv[1], argv[2]);
 }
-
-